// frontend/src/app/http/model/data.ts
/**
 * Account and Transaction API Specification
 * Swagger for Account and Transaction API Specification
 *
 * OpenAPI spec version: v1.1.1
 * Contact: ServiceDesk@openbanking.org.uk
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * Allows setup of an account access request
 */
export interface Data {
    /**
     * Specifies the Open Banking account request types. This is a list of the data clusters being consented by the PSU, and requested for authorisation with the ASPSP.
     */
    Permissions: Array<Data.PermissionsEnum>;
    /**
     * Specified date and time the permissions will expire. If this is not populated, the permissions will be open ended.  All dates in the JSON payloads are represented in ISO 8601 date-time format.  All date-time fields in responses must include the timezone. An example is below: 2017-04-05T10:43:07+00:00
     */
    ExpirationDateTime?: Date;
    /**
     * Specified start date and time for the transaction query period. If this is not populated, the start date will be open ended, and data will be returned from the earliest available transaction.  All dates in the JSON payloads are represented in ISO 8601 date-time format.  All date-time fields in responses must include the timezone. An example is below: 2017-04-05T10:43:07+00:00
     */
    TransactionFromDateTime?: Date;
    /**
     * Specified end date and time for the transaction query period. If this is not populated, the end date will be open ended, and data will be returned to the latest available transaction.  All dates in the JSON payloads are represented in ISO 8601 date-time format.  All date-time fields in responses must include the timezone. An example is below: 2017-04-05T10:43:07+00:00
     */
    TransactionToDateTime?: Date;
}
export namespace Data {
    export type PermissionsEnum = 'ReadAccountsBasic' | 'ReadAccountsDetail' | 'ReadBalances' | 'ReadBeneficiariesBasic' | 'ReadBeneficiariesDetail' | 'ReadDirectDebits' | 'ReadProducts' | 'ReadStandingOrdersBasic' | 'ReadStandingOrdersDetail' | 'ReadTransactionsBasic' | 'ReadTransactionsCredits' | 'ReadTransactionsDebits' | 'ReadTransactionsDetail';
    export const PermissionsEnum = {
        ReadAccountsBasic: 'ReadAccountsBasic' as PermissionsEnum,
        ReadAccountsDetail: 'ReadAccountsDetail' as PermissionsEnum,
        ReadBalances: 'ReadBalances' as PermissionsEnum,
        ReadBeneficiariesBasic: 'ReadBeneficiariesBasic' as PermissionsEnum,
        ReadBeneficiariesDetail: 'ReadBeneficiariesDetail' as PermissionsEnum,
        ReadDirectDebits: 'ReadDirectDebits' as PermissionsEnum,
        ReadProducts: 'ReadProducts' as PermissionsEnum,
        ReadStandingOrdersBasic: 'ReadStandingOrdersBasic' as PermissionsEnum,
        ReadStandingOrdersDetail: 'ReadStandingOrdersDetail' as PermissionsEnum,
        ReadTransactionsBasic: 'ReadTransactionsBasic' as PermissionsEnum,
        ReadTransactionsCredits: 'ReadTransactionsCredits' as PermissionsEnum,
        ReadTransactionsDebits: 'ReadTransactionsDebits' as PermissionsEnum,
        ReadTransactionsDetail: 'ReadTransactionsDetail' as PermissionsEnum
    }
}