// frontend/src/app/http/model/accountRequestPOSTRequest.ts
/**
 * Account and Transaction API Specification
 * Swagger for Account and Transaction API Specification
 *
 * OpenAPI spec version: v1.1.1
 * Contact: ServiceDesk@openbanking.org.uk
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Data } from './data';


/**
 * Allows setup of an account access request
 */
export interface AccountRequestPOSTRequest {
    Data: Data;
    /**
     * The Risk payload is sent by the initiating party to the ASPSP. It is used to specify additional details for risk scoring for Account Info.
     */
    Risk: any;
}