// frontend/src/app/http/model/account.ts
/**
 * Account and Transaction API Specification
 * Swagger for Account and Transaction API Specification
 *
 * OpenAPI spec version: v1.1.1
 * Contact: ServiceDesk@openbanking.org.uk
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Data3Account } from './data3Account';
import { Data3Servicer } from './data3Servicer';


/**
 * Account
 */
export interface Account {
    /**
     * A unique and immutable identifier used to identify the account resource. This identifier has no meaning to the account owner.
     */
    AccountId: string;
    /**
     * Identification of the currency in which the account is held.  Usage: Currency should only be used in case one and the same account number covers several currencies and the initiating party needs to identify which currency needs to be used for settlement on the account.
     */
    Currency: string;
    /**
     * The nickname of the account, assigned by the account owner in order to provide an additional means of identification of the account.
     */
    Nickname?: string;
    Account?: Data3Account;
    Servicer?: Data3Servicer;
}